+++ numerapi/cli.py
@@ -39,17 +39,6 @@ def dataset_url(tournament):
 @cli.command()
 @click.option('--tournament', default=8,
               help='The ID of the tournament, defaults to 8')
-@click.option('--round_num', default=0,
-              help='The selected round, defaults to current round')
-def v1_leaderboard(round_num=0, tournament=8):
-    """Retrieves the leaderboard for the given round."""
-    click.echo(prettify(napi.get_v1_leaderboard(tournament=tournament,
-                                                round_num=round_num)))
-
-
-@cli.command()
-@click.option('--tournament', default=8,
-              help='The ID of the tournament, defaults to 8')
 def competitions(tournament=8):
     """Retrieves information about all competitions"""
     click.echo(prettify(napi.get_competitions(tournament=tournament)))