+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Notable changes to this project.
 
 ## dev
 - improve docstrings and signals example code
+- remove deprecated `get_v1_leaderboard`
 
 ## [2.4.0] - 2021-01-12
 - fix `stake_change` call by adding `tournament` parameter (#32)